FAT fixes
Works in linux and mtools, not yet mac os?

+++ flipsyfat/software/experiment/block.h
@@ -3,7 +3,7 @@
 
 #define BLOCK_SIZE	512
 
-void block_read(uint32_t *buf, uint32_t lba);
-void block_write(uint32_t *buf, uint32_t lba);
+void block_read(uint8_t *buf, uint32_t lba);
+void block_write(uint8_t *buf, uint32_t lba);
 
 #endif // _BLOCK_H
